Restructure project directory.
