+++ src/main/java/zenpad/misc/ButtonFactory.java
@@ -1,4 +1,4 @@
-package zenpad.utils;
+package zenpad.misc;
 
 import java.awt.Insets;
 import java.util.function.Consumer;
