+++ src/main/java/zenpad/misc/DialogUtils.java
@@ -1,4 +1,4 @@
-package zenpad.utils;
+package zenpad.misc;
 
 import javax.swing.Icon;
 import javax.swing.JOptionPane;
